move generic cluster stuff to cluster.c, add conditional compilation of cluster implementation
